+++ scripts/common.rb
@@ -1,3 +1,6 @@
+require "set"
+require "csv"
+
 def convert_terms(upgrade)
   upgrade.gsub! /\[Regenerate\]/i, "[REGENERATE]"
   upgrade.gsub! /\[evade\]/i, "[EVADE]"
@@ -36,18 +39,46 @@ def no_quotes(a)
 end
 
 def convert_line(a)
-  no_quotes(a)
+  if a != nil
+    a.gsub!(/\n+/, " ")
+    a.gsub!(/ +/, " ")
+    a = no_quotes(a)
+  end
+  a
 end
 
 def sanitize_title(title)
   title.gsub(/\W+/, "_")
 end
 
+$external_ids = Set.new()
+
 def make_external_id(setId, title)
-  "#{sanitize_title(title)}_#{setId}".downcase()
+  sanitized_title = sanitize_title(title)
+  external_id = "#{sanitize_title(title)}_#{setId}".downcase()
+  if $external_ids.member?(external_id)
+    letters = 'b'..'z'
+    letters.each do |letter|
+      external_id = "#{sanitize_title(title)}_#{letter}_#{setId}".downcase()
+      unless $external_ids.member?(external_id)
+        break
+      end
+    end
+  end
+  $external_ids.add(external_id)
+  external_id
 end
 
 def set_id_from_expansion(expansion)
   parts = expansion.split (/\s+-\s+/)
   parts[0]
+end
+
+def parse_data(data)
+  rows = []
+  csv = CSV.new(data, {:col_sep => "\t"})
+  csv.each do |row|
+    rows.push(row)
+  end
+  rows
 end