+++ src/DockDataFileLoader.m
@@ -212,6 +212,10 @@ static NSMutableDictionary* createExistingItemsLookup(NSManagedObjectContext* co
                     DockShip* ship = (DockShip*)c;
                     NSString* shipClass =  [d valueForKey: key];
                     [ship updateShipClass: shipClass];
+                } else if ([key isEqualToString: @"ShipClassDetailsId"]) {
+                    DockShip* ship = (DockShip*)c;
+                    NSString* shipClassDetailsId =  [d valueForKey: key];
+                    [ship updateShipClassWithId: shipClassDetailsId];
                 }
             }
 
@@ -552,7 +556,11 @@ static NSString* makeKey(NSString* key)
         @"PlusFourIfNotPredatorClass",
         @"PlusFiveIfNotMirrorUniverse",
         @"OnlyFerengiCaptainFerengiShip",
-        @"OnlyFerengiShip"
+        @"OnlyFerengiShip",
+        @"AllUpgradesMinusOneOnIndepedentShip",
+        @"OnlyFedShipHV4CostPWVP1",
+        @"OnlyBorgShipAndNoMoreThanOnePerShip",
+        @"OnlyNonBorgShipAndNonBorgCaptain"
                                ];
     NSMutableSet* unhandledSpecials = [[NSMutableSet alloc] initWithSet: specials];
     [unhandledSpecials minusSet: [NSSet setWithArray: handledSpecials]];