+++ doc/release-notes.md
@@ -1,3 +1,11 @@
+### Version 1.1.9 build 90 (2014-10-04)
+
+* Add Wave 8 and Avatar of Tomed.
+
+### Version 1.1.8 build 88 (2014-9-26)
+
+* Fix error in when duplicating squads.
+
 ### Version 1.1.7 build 87 (2014-9-18)
 
 * Fix error in Operations Officer.